Windows: Make firstboot a dynamically-generated script
This change allows firstboot to be generated dynamically. Firstly, it involves
an update to rhsrvany. Previously, we created a service called 'rhev-apt', which
ran the firstboot script with rhsrvany. This service was 'uninstalled' when
rhev-apt was installed rhev-apt overwrote it. The new rhsrvany allows us to
specify a service name. We now install a separate service called v2v-firstboot
which runs the firstboot script. The installation of rhev-apt does not affect
it. We use a new function of rhsrvany to uninstall its service at the end of the
firstboot script.

We now write the firstboot script dynamically depending on what needs to go into
it. Currently this is just rhev-apt.

As it is now dynamically generated, the firstboot script is no longer required
in virt-v2v.db or the spec file.